requirements.txt
django<1.10
django-debug-toolbar
gunicorn<20
celery<3.2
redis<3
django-constance<1.3
pytz
vk